[SG-52] move: navigator 패키지 분리
