+++ core/navigator/src/main/kotlin/com/captures2024/soongan/core/navigator/screen/main/profile/ProfileNavigator.kt
@@ -1,4 +1,4 @@
-package com.captures2024.soongan.core.navigator.screen.main
+package com.captures2024.soongan.core.navigator.screen.main.profile
 
 import kotlinx.serialization.Serializable
