+++ feature/awards/src/main/kotlin/com/captures2024/soongan/feature/awards/navigation/AwardsNavigation.kt
@@ -2,7 +2,7 @@ package com.captures2024.soongan.feature.awards.navigation
 
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.composable
-import com.captures2024.soongan.core.navigator.screen.main.AwardsNavigator
+import com.captures2024.soongan.core.navigator.screen.main.awards.AwardsNavigator
 import com.captures2024.soongan.feature.awards.route.AwardsRoute
 
 fun NavGraphBuilder.awards() {
