+++ feature/feed/src/main/kotlin/com/captures2024/soongan/feature/feed/navigation/FeedNavigation.kt
@@ -2,7 +2,7 @@ package com.captures2024.soongan.feature.feed.navigation
 
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.composable
-import com.captures2024.soongan.core.navigator.screen.main.FeedNavigator
+import com.captures2024.soongan.core.navigator.screen.main.feed.FeedNavigator
 import com.captures2024.soongan.feature.feed.route.FeedRoute
 
 fun NavGraphBuilder.feed() {
