+++ feature/home/src/main/kotlin/com/captures2024/soongan/feature/home/navigation/HomeNavigation.kt
@@ -3,11 +3,11 @@ package com.captures2024.soongan.feature.home.navigation
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.composable
 import com.captures2024.soongan.core.model.UserPost
-import com.captures2024.soongan.core.navigator.screen.main.HomeGalleryNavigator
-import com.captures2024.soongan.core.navigator.screen.main.HomeNavigator
-import com.captures2024.soongan.core.navigator.screen.main.HomePostNavigator
-import com.captures2024.soongan.core.navigator.screen.main.HomePostPhotoNavigator
-import com.captures2024.soongan.core.navigator.screen.main.RegistrationPostNavigator
+import com.captures2024.soongan.core.navigator.screen.main.home.HomeGalleryNavigator
+import com.captures2024.soongan.core.navigator.screen.main.home.HomeNavigator
+import com.captures2024.soongan.core.navigator.screen.main.home.HomePostNavigator
+import com.captures2024.soongan.core.navigator.screen.main.home.HomePostPhotoNavigator
+import com.captures2024.soongan.core.navigator.screen.main.home.RegistrationPostNavigator
 import com.captures2024.soongan.feature.home.route.HomeGalleryRoute
 import com.captures2024.soongan.feature.home.route.HomePostPhotoRoute
 import com.captures2024.soongan.feature.home.route.HomePostRoute
