+++ feature/welcome/src/main/kotlin/com/captures2024/soongan/feature/welcome/route/WelcomeRoute.kt
@@ -5,7 +5,7 @@ import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.navigation.NavOptions
 import androidx.navigation.navOptions
-import com.captures2024.soongan.core.navigator.screen.main.WelcomeNavigator
+import com.captures2024.soongan.core.navigator.screen.main.welcome.WelcomeNavigator
 import com.captures2024.soongan.feature.welcome.WelcomeUiState
 import com.captures2024.soongan.feature.welcome.WelcomeViewModel
 import com.captures2024.soongan.feature.welcome.ui.WelcomeScreen
